Treat MMFSR, BFSR and UFSR as part of CFSR in SCB
Those are actually parts of the CFSR register. And since this tool
doesn't care if fields are overlapping, they can be treted and aligned
as fields instead, which makes printouts nicer

This however needs fields to be forced to be written, since they are
registers and therefore should always be printed out.

As side effect, add the /a modifier so all registers can be forced to be
written out